add pagination feat

+++ bot/database/models.py
@@ -1,5 +1,5 @@
 from typing import Optional
-from sqlalchemy import ForeignKey, BigInteger, Column
+from sqlalchemy import ForeignKey, BIGINT, Column
 from sqlalchemy.orm import mapped_column, Mapped
 from database.database import Base
 
@@ -11,7 +11,7 @@ class Questions(Base):
 
     text: Mapped[str]
 
-    chat_id = Column(BigInteger)
+    chat_id = Column(BIGINT)
 
     admin_id: Mapped[Optional[int]] = mapped_column(ForeignKey("admins.id"))
 
@@ -21,6 +21,6 @@ class Admins(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
 
-    chat_id = Column(BigInteger)
+    chat_id = Column(BIGINT)
 
     quesiton: Mapped[Optional[int]] = mapped_column(ForeignKey("questions.id"))
